# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AVITO API",
        "description": "API endpoints for AVITO",
        "contact": {
            "email": "pashaboga4ev@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://0.0.0.0:80/api",
            "description": "Base server for development"
        }
    ],
    "paths": {
...
        "/feature": {
            "get": {
                "tags": [
                    "Avito Feature"
                ],
                "summary": "Get list of features",
                "operationId": "featuresList",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Токен админа",
                        "schema": {
                            "type": "string",
                            "example": "admin_token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/FeatureCollectionResponse"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    }
                }
            },
            "post": {
                "tags": [
                    "Avito Feature"
                ],
                "summary": "Create feature",
                "operationId": "createFeature",
                "parameters": [
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Токен админа",
                        "schema": {
                            "type": "string",
                            "example": "admin_token"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/CreateFeatureRequestBody"
                },
                "responses": {
                    "201": {
                        "$ref": "#/components/responses/FeatureResponse"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "422": {
                        "$ref": "#/components/responses/InvalidDataResponse"
                    }
                }
            }
        },
        "/feature/{id}": {
            "get": {
                "tags": [
                    "Avito Feature"
                ],
                "summary": "Get feature by id",
                "operationId": "getFeature",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Токен админа",
                        "schema": {
                            "type": "string",
                            "example": "admin_token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/FeatureResponse"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    }
                }
            },
            "put": {
                "tags": [
                    "Avito Feature"
                ],
                "summary": "Update feature",
                "operationId": "updateFeature",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Токен админа",
                        "schema": {
                            "type": "string",
                            "example": "admin_token"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/UpdateFeatureRequestBody"
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/FeatureResponse"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    },
                    "422": {
                        "$ref": "#/components/responses/InvalidDataResponse"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Avito Feature"
                ],
                "summary": "Delete feature",
                "operationId": "deleteFeature",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "token",
                        "in": "header",
                        "description": "Токен админа",
                        "schema": {
                            "type": "string",
                            "example": "admin_token"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/SuccessDeletedResponse"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedResponse"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFoundResponse"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateFeatureRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "feature name"
                    }
                },
                "type": "object"
            },
            "UpdateFeatureRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "feature name"
                    }
                },
                "type": "object"
            },
...
            "FeatureCollection": {
                "required": [
                    "data"
                ],
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/ResourceCollection"
                    },
                    {
                        "properties": {
                            "data": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/FeatureResource"
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "FeatureResource": {
                "required": [
                    "feature_id",
                    "name",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "feature_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "feature name"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ResourceCollection": {
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "meta": {
                        "properties": {
                            "current_page": {
                                "type": "number",
                                "example": 2
                            },
                            "from": {
                                "type": "number",
                                "example": 1
                            },
                            "last_page": {
                                "type": "number",
                                "example": 8
                            },
                            "links": {
                                "type": "object"
                            },
                            "path": {
                                "type": "string",
                                "example": "http://0.0.0.0/api/v1/object/list?page=2"
                            },
                            "per_page": {
                                "type": "number",
                                "example": 1
                            },
                            "to": {
                                "type": "number",
                                "example": 1
                            },
                            "total": {
                                "type": "number",
                                "example": 1
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "InvalidData": {
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Any error message"
                    }
                },
                "type": "object"
            },
            "NotFound": {
                "required": [
                    "error"
                ],
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "Not Found"
                    }
                },
                "type": "object"
            },
...
            "Unauthorized": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthorized"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "FeatureCollectionResponse": {
                "description": "Success",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/FeatureCollection"
                        }
                    }
                }
            },
            "FeatureResponse": {
                "description": "Success",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/FeatureResource"
                        }
                    }
                }
            },
...
            "InvalidDataResponse": {
                "description": "Некорректные данные",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/InvalidData"
                        }
                    }
                }
            },
            "NotFoundResponse": {
                "description": "Not Found",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/NotFound"
                        }
                    }
                }
            },
...
            "SuccessDeletedResponse": {
                "description": "Success Deleted"
            },
            "UnauthorizedResponse": {
                "description": "Пользователь не авторизован",
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Unauthorized"
                        }
                    }
                }
            },
...
        },
        "requestBodies": {
...
            "CreateFeatureRequestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/CreateFeatureRequest"
                        }
                    }
                }
            },
            "UpdateFeatureRequestBody": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/UpdateFeatureRequest"
                        }
                    }
                }
            },
...
        }
    },
...
}